fix: 将课表分割线宽度单位从像素改为dp
修改 AppSettings 模型中的 dividerWidth 字段为 dividerWidthDp，类型从 Int 改为 Float，以明确使用 dp 作为单位。同步更新 SettingsRepository、SettingsRepositoryImpl、SettingsViewModel 和 TimetableComponents 中的相关方法及存储逻辑，确保数据层、业务层和 UI 层单位一致。在 SettingsScreen 中调整了部分 UI 文本和布局细节。

+++ core/domain/src/main/java/com/dawncourse/core/domain/repository/SettingsRepository.kt
@@ -38,7 +38,7 @@ interface SettingsRepository {
     suspend fun setFontStyle(style: AppFontStyle)
     
     suspend fun setDividerType(type: com.dawncourse.core.domain.model.DividerType)
-    suspend fun setDividerWidth(width: Int)
+    suspend fun setDividerWidth(width: Float)
     suspend fun setDividerColor(color: String)
     suspend fun setDividerAlpha(alpha: Float)
